refactor(shared): reorganizar extensão de DI e endurecer cobertura do accessor
Acolhe sugestões S1, S2 e S4 do review do PR #102.

S1 — separar extensão de DI do folder Middleware/:
CorrelationIdServiceCollectionExtensions sai de
src/shared/Infrastructure.Common/Middleware/ para
src/shared/Infrastructure.Common/DependencyInjection/. A extensão é
responsabilidade de wiring, não middleware — separar o folder clarifica
o papel e mantém simetria com o padrão usual do ecossistema .NET. Os
using de Program.cs (Selecao e Ingresso) passam a importar também
o namespace DependencyInjection.

S2 — teste direto da invariante de singleton compartilhada:
Novo arquivo DependencyInjection/CorrelationIdServiceCollectionExtensionsTests.cs
com dois testes: (a) ICorrelationIdAccessor e ICorrelationIdWriter
resolvem para a MESMA instância de CorrelationIdAccessor, (b)
AddCorrelationIdAccessor(null) lança ArgumentNullException. Protege
contra regressão se alguém substituir as factories por
AddSingleton<I, Impl>() separado — que resolveria duas instâncias
distintas e quebraria silenciosamente o fluxo de correlação.

S4 — comentário em InvokeAsync_DurantePipeline_AccessorDeveRetornarMesmoValor:
Três linhas explicando por que new CorrelationIdAccessor() no teste é
seguro (AsyncLocal estático) enquanto em produção a instância vem via
DI como Singleton. Evita leituras confusas no futuro.

Follow-up S3 (integration tests via WebApplicationFactory) ficará em
issue separada — exige decisão sobre a infra dos projetos
*.Integration.Tests já existentes.

Build: 0 warnings / 0 errors.
Testes: 25/25 aprovados (era 23/23 + 2 novos de S2).
